requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
google-generativeai==0.8.3
google-auth==2.40.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
PyPDF2==3.0.1
pdfplumber==0.11.4
pillow==10.4.0
python-dotenv==1.0.1
werkzeug==3.0.1
reportlab==4.4.1
authlib==1.3.0
requests==2.32.3